implement multi reader solution
A list of open file is maintained. Each file has its own buffer and a
write operation goes and writes the the queue of each files.
If one queue is full, no queue is updated at all.
The access to the list of open files is protected by a mutex.
If a file is opened write only, it won't copy to its own private
descriptor buffer. If it does, it is possible to block everything by
giving a very big string to the device via echo. The consumer will
consume the very first writen character but the queue opened by the
writer won't be emptied and hence creates a deadlock