Fix CI regression
Add missing clang-12 dependencies which prevented the flags to be
properly detected.

Add missing libclang-cpp12-dev to allow Easymock to be compiled.

Updated to the greatest and latest EasyMock version.

Fix unit test because list_for_each invokes now a function which needs
to be mocked.

Add the possibility to specify the clang command to detect the flag to
be able to run the test locally